Simplify array_container_union
The code was simplified and a little optimization was made such that the
smallest array is passed first.

The benchmark is a tiny bit faster, but I believe this is due to the
construction of arrays and the order the comparators in the loops are
verified. The worst case would be a small array with only one element
that is bigger than all elements of the second (bigger) array.