Make C files build w/C++ compiler, cleanups
The Roaring sources *nearly* built with C++.  The only barriers were
some inconsistencies in `inline` vs `static inline`...and use of tagged
structure initialization for making `rle16_t` pairs.

This adds a MAKE_RLE16() macro that uses tags in C, but leaves them off
in C++.  Using the macro makes the callsites more readable, and
consistent.

`extern "C" {...}` brackets the *contents* of headers.  Cases where it
spanned inclusion of other headers (like `<assert.h>` or `<stdbool.h`)
are fixed.  Having all the headers compliant with extern "C" permits
things like C++ test code to be able to reach in at any level of
implementation (e.g. to do a comparison of the behavior of a low-level
container against standard C++ library types like `std::set`).

A new CMake option ROARING_BUILD_C_AS_CPP is added.  This instructs
it to treat the files in the library itself that end in `.c` as if
they were C++.  Being able to use a C++ compiler to build C codebases
is a useful test for catching certain errors--and it also offers a
rich set of static analysis capabilities (e.g. with `<type_traits>`)