roaring_checked.hh differential check vs std::set
This creates a C++ test class which maintains a parallel state to a
roaring bitmap with a `std::set` of 32-bit integers.  On each method
call, the test class ensures that the Roaring bitmap gives the same
answer that the std::set would.

The class is in the file roaring_checked.hh, and has a superset of the
methods as the exported C++ class in roaring.hh.  It even has the same
name of `Roaring`, but under the `doublechecked` namespace.  This means
tests written to use the class are also testing the C++ class, and that
any ordinary codebase using the C++ class could act as a test if it
wanted to.

For operations like intersect/xor/union, the test class has a custom
implementation of things that have implementations in `<algorithm>`.
The reason this is done is partly because those algorithms are not
available in in-place forms--so they would use extra space.  But also
some of these algorithms like `set_intersection()` are not available
in pre-C++17 compilers.  The algorithms aren't terribly long, and
since std::set is sorted they just involve bumping pointers linearly
along the two sets with handling for cases where they match.

To exercise this code, a randomizing test was added that can run
for an indefinite number of cycles.  It is added to the unit tests and
run for a small number of operations, to make sure it stays compiling.
Future improvements to this test could use it as part of a coverage
guided fuzz test, that used something like LLVM's libFuzzer to make
sure the tests were actually exercising all the various branches in
the C bitset implementation.