Separate into ::api and ::internal namespaces
A typical C++ client would want to use the C++ interface definitions
and have nothing else in the global namespace.  This would suggest
that `using namespace roaring;` would not bring in C APIs, but only
C++ classes and utilities.

As it would be nice for testing or other purposes to have a way to get
access to the same APIs a C client (should have), this puts those under
the namespace `roaring::api`.

Everything that doesn't fit into these categories is put under the
namespace `roaring::internal`...with the exception of the contents
of configreport.h which are put under `roaring::misc`.

This tries to follow the C++ best practice of not putting `using`
statements in header files, where possible.  But since it needs to
compile as C there have to be a few exceptions.