Add C++ Namespaces (retry to trigger Appveyor) (#253)
* Make C files build w/C++ compiler, cleanups

The Roaring sources *nearly* built with C++.  The only barriers were
some inconsistencies in `inline` vs `static inline`...and use of tagged
structure initialization for making `rle16_t` pairs.

This adds a MAKE_RLE16() macro that uses tags in C, but leaves them off
in C++.  Using the macro makes the callsites more readable, and
consistent.

`extern "C" {...}` brackets the *contents* of headers.  Cases where it
spanned inclusion of other headers (like `<assert.h>` or `<stdbool.h`)
are fixed.  Having all the headers compliant with extern "C" permits
things like C++ test code to be able to reach in at any level of
implementation (e.g. to do a comparison of the behavior of a low-level
container against standard C++ library types like `std::set`).

A new CMake option ROARING_BUILD_C_AS_CPP is added.  This instructs
it to treat the files in the library itself that end in `.c` as if
they were C++.  Being able to use a C++ compiler to build C codebases
is a useful test for catching certain errors--and it also offers a
rich set of static analysis capabilities (e.g. with `<type_traits>`)

* First Step in Namespacing: One Big Namespace

This puts everything (API public and internal) into a single
giant `roaring` namespace when building as C++.

* Separate into ::api and ::internal namespaces

A typical C++ client would want to use the C++ interface definitions
and have nothing else in the global namespace.  This would suggest
that `using namespace roaring;` would not bring in C APIs, but only
C++ classes and utilities.

As it would be nice for testing or other purposes to have a way to get
access to the same APIs a C client (should have), this puts those under
the namespace `roaring::api`.

Everything that doesn't fit into these categories is put under the
namespace `roaring::internal`...with the exception of the contents
of configreport.h which are put under `roaring::misc`.

This tries to follow the C++ best practice of not putting `using`
statements in header files, where possible.  But since it needs to
compile as C there have to be a few exceptions.