Additional namespacing in roaring.hh
The unit test for the C++ class had a `using roaring::api;` inclusion to
access the C functions.  As a consequence, it did not catch missing api::
namespacing on `roaring_bitmap_t` and the iterator types.

Because roaring_bitmap_t is used several times, this creates a local
typedef alias...to avoid a `using` statement in a header file.