Add namespace to generated amalgamation_demo.cpp
Users of the C++ class must now use Roaring from namespace `roaring`
The file `amalgamation_demo.cpp` is generated by a shell script, and
the script wasn't modified to use this namespacing.

This also exposed that the 64-bit class was dependent on the api::
namespace inclusion from the unit test.  When that namespace is not
included, it needs to qualify the `roaring_bitmap_t` and iterator
datatypes.