Unify ra_copy() and ra_overwrite()
Finding a problem in ra_overwrite() that had been worked around in
ra_copy() drew attention to how similar the two routines were (and
that a difference was only a bug).

The semantics were that ra_copy() assumed it was writing into
uninitialized memory while ra_overwrite() was willing to reuse the
existing capacity of a roaring_array_t if it was big enough.

This reframes the roaring_bitmap_copy() routine as initializing a
list of containers of the appropriate size (so it won't need a
realloc_array()) and then calling ra_overwrite().  ra_copy() is
then deleted.