Remove dependence of C++ class on roaring_array.h
This takes another step to reducing the export surface for the API,
by adding a missing routine needed by C++ to initialize a roaring
bitmap structure in client-controlled memory.

Previously the C++ class had to include roaring_array.h and use
ra_init(), otherwise the only way it could create a bitmap would
be through a malloc()'d handle.  roaring_bitmap_init_with_capacity()
is added to the API to provide this.

(roaring_bitmap_init_cleared() is created as an inline that calls
roaring_bitmap_init_capacity() with 0, to show that by contract in
the API that these are not semantically different.  Since it cuts
down on variation, the change is applied to roaring_bitmap_create()
as well, as an inline call to roaring_bitmap_create_with_capacity())

The other features used by the C++ class were ra_copy(), which
already had an export suitable for the functionality in
roaring_bitmap_overwrite() (rebased onto a bugfix for that). Also
ra_clear(), effectively available as roaring_bitmap_clear()

With this change, roaring_array.h and containers.h can be left out
of the usage of the C++ API as well; leading to *many* fewer exports.