hide containers.h/roaring_array.h from public api
The function roaring_bitmap_contains() in the C API was declared as
inline, pulling in a considerable set of dependencies.  Usage of the
container_contains() function brought in definitions for
bitset_container_t, array_container_t, run_container_t, as well as
all the needed support functions.

While inlining can give considerable advantages, the definition of
this particular public export as inline causes collateral damage
that is unlikely to move the needle much to client performance.

This migrates the implementation of roaring_bitmap_contains() into
roaring.c, and moves the `roaring_array_t` from roaring_array.h
into roaring_types.h.  This then allows containers.h to be removed
from the public API.

Code such as tests that wants to use the internal functions thus
must include the internal headers to do so.  (Using the internal API
to reuse code is also an axis that some clients might take in
pursuit of greater performance, if they were willing to deal with
internals changing.)