Use private/public separation in amalgamation.sh
Though the export dependencies for roaring.h and roaring.hh were
reduced, the amalgamation script put all the header files into
the aggregated roaring.h -- whether they were public or private.

This separates the headers out, so that the private ones are
packed into the .C file instead of the .H file.

The .HH file is also adjusted to use the #define guard that keeps
the C API from being put in the global namespace by default when
it re-adds the `#include roaring.h` that it stripped out.

Minor changes to reduce redundancy are included--factoring the
timestamp and generated file warning into a function, an omitting
the original-file-location comments when including the license.