Sync XOR submethods to container_xor parameters
This is a simplification which is suggested for all operations,
but is demonstrated first for XOR.

It makes the function signatures for the subdispatch of xor
for container pairs match the signature for the higer-level
dispatch.

This means there is no need for an encoding scheme that returns
booleans that have to be translated into container types.  The
address of the container type being fulfilled can be passed
through, and not written to at all if the container type is
not altered.