Use a macro for formulaic default inplace XORs
Not all pairings of containers had an inplace algorithm
for XOR written.  So they would just delegate to the normal
XOR and free the input.

This makes it clearer which methods have not been optimized,
by using a macro to generate the functions.  This also means
that any systemic change to one such wrapper (such as out of
memory handling) would be reflected in the others.