Make C Tests Compilable in C++ Mode
The CMake build switch ROARING_BUILD_C_AS_CPP was only being
applied to the main files, and not the tests.  This commit adds
another option for the test files: ROARING_BUILD_C_TESTS_AS_CPP

(Having a separate setting is beneficial, in order to make sure
a C++ built library can be used with C clients...and vice versa)

With this change, there are no `void*` references to containers
or raw casts.  This gives compile time errors when aspects of
the internal interfaces change that would be more difficult to
debug otherwise.

Because there is no compatible way to declare a function across
C and C++ that doesn't name a `void**` argument, this required
making the test functions use a macro: DEFINE_TEST()