Tunnel boolean result from inplace XOR
This continues using XOR as an example for how dispatch might
be rethought, in order to bubble up a signal that an inplace
XOR failed.  The same method would apply to other operators.

Because inplace xor is a performance-oriented mechanism, it
would not be worth it to keep a log of changes to be able to
roll back the input atomically if an out-of-memory error
occurred halfway through the operation.  Therefore when a
false result is returned, the input being mutated is put into
an invalid state--with all auxiliary memory released.

For legacy compatibility and ease of use, the original xor
operations are implemented as inline functions that call an
overrideable failure handler in the case of a problem.  The new
operators which report failure are given new names.