Inplace handling of shared container in dispatcher
The higher level routine wasn't passing shared containers to the
container_ixor() or container_xor() functions.  This made it not clear
if they would handle shared containers (one asserted it did not, while
the other unwrapped the never-passed shared version).

This moves it into the dispatcher for clarity, and also to cut down
to just one dispatcher call from the container enumerator routine.