Conform Lazy XOR to dispatcher return convention
This continues the push to homogenize the dispatcher functions with
the container combiners for XOR, this time with the lazy functions.

Due to a case always returning a run container and being reused by
the non-lazy code, it returns a run_container_t* instead of a
container_t*; though this can still be used in places requiring a
container_t* result.

A few lazy functions were willing to accept an already allocated
container.  This created some confusion about whether that container
could be the same as the inputs, and meant callsites were passing
in empty allocated containers.  The only usage of this was in the
test code, so it is removed.