CONTAINER_EMPTY magic pointer to reduce dispatches
The loops for processing containers in XOR would never put an empty
container into the list.  To maintain this invariant, they would do
a dispatch on the container to ask if it had nonzero cardinality.
This involves a switch() on the container type.  Then if it was zero,
another switch() would be needed to free it.

This pushes responsibility for returning a special CONTAINER_EMPTY
magic pointer value onto the combining functions.  Those functions
already know what kind of container they have in hand, and so it is
more expedient for them to empty and free them.

Helper routines `array_result()`, `run_result()`, and `bitset_result()`
combine the task of assigning the result container type along with
doing the conversion to the empty state.  However, if the function
knew it was going to return an empty result sooner it could do that
as well; potentially bypassing unnecessary container creation.