Rename bitset words array "array" => "words"
Having the bitset container use a unique name for its words array
helps significantly in code understandability, especially when
working with a mixture of array containers and bitset containers.
There is no useful polymorphism from the shared name to be lost.

This also standardizes the parameter name for extracted words
arrays passed to `words` as well (variants were `array`, `bitmap`
and `bitset`).