Simplistic runtime debug hooking for malloc/free
In order to inject failures into malloc, one needs to link against a
library that overrides the standard malloc() definition or go through a
custom level of indirection.

This is a simple implementation of the latter for debug builds, which
allows runtime hooks to override the allocation functions.  This is made
more complicated since both malloc() and realloc() are used, as well as
an aligned form of allocation and freeing for the bitset words.