Out-of-memory handling hit by XOR stress test
This adds handling for out of memory conditions hit during the
run of the cpp_random_unit for XOR.

Some cases have a container in hand which could serve as the
data but does not fit the invariants, so it errors on the allocation
of the more compact form...rather than return the too big container
it already has.

Run container XOR was only being called with a freshly created
container, and this changes it to return a new container of the
correct size rather than have an unused boolean protocol for a
two-step process.