Use platform independent random for random tests
This moves the "our_rand()" function into the common test file.
It existed to provide a platform-independent pattern for random number
generatig.  This uses that in the cpp-random-unit as well.

Because it declares state, the file `test.h` is renamed to
`test.inc` to show it is not just prototypes and macros.