Rename ra_mark_corrupt to ra_mark_indeterminate
The first cut of setting the bitmap into a state after an inplace
operation allocation failure didn't account for the idea that there
would need to be a way to tell if the structure was in this state.  So
the memory was trashed with no flag to mark the condition.  The function
had the somewhat charged name "ra_mark_corrupt"

It quickly turned out that a flag was needed for roaring_bitmap_clear()
to know whether to clear the containers in a structure or just set the
pointer to NULL.  This got the softer name ROARING_FLAG_INDETERMINATE.
But the function itself was not renamed.

To avoid triggering people over something that looks "non-standard",
this emphasizes the setting of the in-band flag, and reinforces the
by-contract rules of not calling any APIs besides roaring_bitmap_init()
or roaring_bitmap_clear().  The pointer should not be read, hence what
value it contains should not be relevant...so it's best to use any
value that is not NULL (as that could lead to interpretation of the
bitmap as simply empty).