Allow users to define their own TERMINATE method.
If the user wants to provide a specific termination method they can
define a ROARING_TERMINATE macro taking a single const char* argument.

Whatever this macro expands to must call some function that never
returns.

The default for environments where exceptions are enabled is to throw
std::runtime_error().  For environments where exceptions are disabled
the default is to invoke std::terminate().