Reserve `r` name for `roaring_bitmap_t` variables
A few locations used the single letter `r` for a run container.
The likely preferable name is `rc`, especially in light of the
near term need to pass roaring_bitmap_t* frequently to many routines.

This adjusts those locations, and goes ahead and adjusts `a` for
array container and `b` for bitset container as well in the few places
that used the notation.