Tidy `ra` vs `r` usages, roaring.h comment cleanup
The "ra" name has been used for roaring_array_t.  With a potential
increase in the number of systemic references to the API-visible idea
of a roaring_bitmap_t, the distinction of `r` for that idea and
`ra` for the roaring_array_t is more important.

This adjusts all the instances of `roaring_bitmap_t *ra`, in the API
and elsewhere.  In the API, most all the roaring_bitmap_t* parameters
are standardized to `r` (instead of ra, x, bmp...)

Inside roaring.c, the repetitive dereference of high_low_container is
replaced with extracting an `ra` from the `r` and then using that.
This increases readability.

Since many roaring.h comments had to be edited to account for the new
names, this goes ahead and takes a crack at fixing up those comments
as a whole...though it's something that could use more work.

Routines in roaring.c that didn't use `ra` as a name, but used names
like `x1` instead of `r1` are not adjusted by this commit at this time.