portability.h: add an build flag to suppress warnings
When building CRoaring without support for atomics (for instance, in a
pre-C11 environment where <stdatomic.h> is not part of the language
specification), we get a compile-time message that copy-on-write bitmaps
are unavailable.

Having a warning message is useful as a default behavior, since it can
prevent surprises for CRoaring users who expect their bitmaps to be
thread-safe, but aren't for whatever reason.

But in environments where we know that <stdatomic.h> is unavailable, or
we are single-threaded, the message is noise, since we know ahead of
time that CRoaring's bitmaps won't be threadsafe (and are OK with that).

Provide an opt-out build knob (CROARING_SILENT_BUILD) to suppress this
warning for applications that wish to do so.

Signed-off-by: Taylor Blau <me@ttaylorr.com>