fix(art): node48 children can have holes
This matches the [java implementation][1] in Roaring.

When removing from a node48, the child is simply set to NULL. For example, if
the state is (ignoring that it actually needs at least 17 items to not be a
node16):

```
keys: {[0]=0, [1]=1}
children: {[0]=child1, [1]=child2}
count: 2
```

After removing key 0, the state will be:

```
keys: {[0]=48, [1]=1}
children: {[0]=NULL, [1]=child2}
count: 1
```

If we then try to insert at key 3, the item at `children[count]` is already in
use: we have to find and reuse the 0th child.

This also has some fixes for `art_printf`:

* When printing keys, don't cast to char, if char is signed, when integer
  promoted, a 0xFF char will be printed as `FFFFFFFF`
* Correctly output keys for node48
* Include child index for node48
* Correct child output for node48

[1]: https://github.com/RoaringBitmap/RoaringBitmap/blob/1ed0242d1f298cf93d3666a754eb4bb4c5180c4c/RoaringBitmap/src/main/java/org/roaringbitmap/art/Node48.java#L181-L187