r64 iterator: allow moving from beyond the start / end (#578)
Previously calling `roaring64_iterator_previous` after
`roaring64_iterator_advance` returned false would not move the iterator. With
this commit, this behavior and its inverse are both possible.

I chose to implement this by adding a new field in the iterator to indicate
saturation. Another option would be to allow the ART iterator to allow moving
from beyond the start / end, but that would require changing the index to a
signed integer and thus take up more space.