Run clang-format on the repo, add a github action to check formatting (#577)
* Add clang-format directives to order includes

* Run clang-format.sh on the entire repo

* Add previous formatting commit to ignored revisions

This allows anyone to ignore the commit with:
   git blame --ignore-revs-file .git-blame-ignore-revs

Or to always use the ignore-revs file:
    git config blame.ignoreRevsFile .git-blame-ignore-revs

* Fix some include orders

* Add a github action to check clang-format

* Reference formatting in README and PR template

* Use .clang-format-ignore for github action

The github action expects a .clang-format-ignore file in order to exclude files,
so I've renamed it. It also expects the entries to match the full path as far as
I can tell, so I've switched them to be regexes rather than plain filenames.