Sort free lists in art_shrink_to_fit
This avoids a bug in the following scenario:
  art->leaves = [2,0,x]
  art->first_free[leaf_type] = 1

Where `2` and `0` are pointers to the next free index, and `x` is an occupied
leaf. In this case, if `art_shrink_to_fit` was called, then we would have the
following result:
  art->leaves = [2,x,0]
  art->first_free[leaf_type] = 0

This is not fully shrunken, and therefore wrong. Sorting the free indices fixes
this scenario. Before `art_shrink_to_fit`:
  art->leaves = [1,2,x]
  art->first_free[leaf_type] = 0

After `art_shrink_to_fit`:
  art->leaves = [x,2,3]
  art->first_free[leaf_type] = 1