fix: avoid out of bounds index in PyRoaringBitMap124 test (#701)
There are only 28 bytes of data, but a size of 36 was being passed.
Instead, make this impossible: convert to a stack array, and use sizeof
rather than a magic number to pass the length of data.

It isn't clear where the test data came from: the linked issue includes
a base64 input of `OjAAAAEAAAAAAAkAEAAAADIAMwA0ADUANgA3ADgAOgA7ADwA`,
which _does_ expand to 36 bytes, so additionally, this replaces the test
data with the decoded value of the data from the linked issue instead.

This was (correctly) failing in ASAN when we immediately read past the
end of the string literal.