Run CI against illumos and make the tests pass.
We only want to use the `restrict` keyword in C, not in C++. Under
linux, darwin, etc., we can check for `__STDC_VERSION__` to verify that
we're compiling C and not C++. However, illumos defines
`__STDC_VERSION__` whether we're compiling C *or* C++. To ensure that we
don't use the `restrict` keyword in C++ on illumos, this patch adds a
check on `__cplusplus` as well, and skips `restrict` if it's defined.

This patch also adds a test workflow for illumos to avoid future
regressions.